Changed the dtypes of classic control envs to float32

+++ gym/envs/classic_control/pendulum.py
@@ -59,7 +59,7 @@ class PendulumEnv(gym.Env):
 
     def _get_obs(self):
         theta, thetadot = self.state
-        return np.array([np.cos(theta), np.sin(theta), thetadot])
+        return np.array([np.cos(theta), np.sin(theta), thetadot], dtype=np.float32)
 
     def render(self, mode="human"):
         if self.viewer is None:
